Move to shaka.dash.JsonUtils
