+++ lib/dash/dash_parser.js
@@ -10,6 +10,7 @@ goog.require('goog.asserts');
 goog.require('goog.Uri');
 goog.require('shaka.abr.Ewma');
 goog.require('shaka.dash.ContentProtection');
+goog.require('shaka.dash.JsonUtils');
 goog.require('shaka.dash.MpdUtils');
 goog.require('shaka.dash.SegmentBase');
 goog.require('shaka.dash.SegmentList');
@@ -477,7 +478,7 @@ shaka.dash.DashParser = class {
       const jsonString = shaka.util.StringUtils.fromBytesAutoDetect(data);
       try {
         const json = /** @type {!Object} */ (JSON.parse(jsonString));
-        const xmlString = shaka.dash.MpdUtils.jsonToMpd(json);
+        const xmlString = shaka.dash.JsonUtils.jsonToMpd(json);
         mpd = shaka.util.TXml.parseXmlString(xmlString, rootElement);
       } catch (e) {
         throw new shaka.util.Error(
